+++ configuration/__init__.py
@@ -1,0 +1,7 @@
+from .configuration import (  # noqa: F401
+    Configuration,
+    ConfigurationFile,
+    ConfigurationItem,
+    ConfigurationManager,
+    ConfigurationProject,
+)
