+++ exceptions/exceptions.py
@@ -1,0 +1,22 @@
+class ProjectManConfigTypeInvalidError(Exception):
+    pass
+
+
+class FieldNotInConfigurationFieldsError(Exception):
+    pass
+
+
+class FieldValidatorNotExistsError(Exception):
+    pass
+
+
+class ProjectManValidationError(Exception):
+    pass
+
+
+class ProjectManInvalidJsonFileError(Exception):
+    pass
+
+
+class GithubObjectNotFoundError(Exception):
+    pass
