+++ mockers/__init__.py
@@ -1,0 +1,1 @@
+from .mockers import ConfigurationMocker, GithubMocker  # noqa: F401
