+++ parsers/__init__.py
@@ -1,0 +1,1 @@
+from .json_parser import JsonParser  # noqa: F401
