+++ providers/__init__.py
@@ -1,0 +1,1 @@
+from .github import GithubProvider  # noqa: F401
