+++ test_field_validator.py
@@ -1,24 +1,17 @@
 import pytest
 
-from main import (
-    ALL,
-    CONFIGURATION_ITEM_ACCEPTED_TYPES,
-    ISSUES,
-    PULL_REQUESTS,
-    FieldValidator,
-    ProjectManValidationError,
-)
+from exceptions import ProjectManValidationError
+from mockers import ConfigurationMocker
+from utils import ALL, CONFIGURATION_ITEM_ACCEPTED_TYPES, ISSUES, PULL_REQUESTS
+from validators import FieldValidator
 
 field_validator = FieldValidator()
-json_dict = {
-    "labels": ["bug"],
-    "assignees": ["thepetk"],
-    "reviewers": ["thepetk"],
-    "milestones": ["12/12/2023"],
-    "created_on": "dmdmd",
-    "last_updated_on": "fefde",
-    "closed_on": "rfcrsd",
-}
+mocker = ConfigurationMocker(
+    labels=["bug"],
+    assignees=["someone", "!another"],
+    reviewers=["reviewer", "!noreviewer"],
+    milestones=["important", "!unimportant"],
+)
 
 
 def test_field_validator_validate_success():
@@ -26,18 +19,26 @@ def test_field_validator_validate_success():
     field_validator.is_one_of = CONFIGURATION_ITEM_ACCEPTED_TYPES
     field_validator.default = ALL
     for value in values:
-        json_dict["type"] = value
-        assert field_validator.validate("type", json_dict["type"]) == json_dict["type"]
-    json_dict["type"] = "None"
-    assert field_validator.validate("type", json_dict["type"]) == ALL
+        mocker.config_dict["type"] = value
+        assert (
+            field_validator.validate("type", mocker.config_dict["type"])
+            == mocker.config_dict["type"]
+        )
+    mocker.config_dict["type"] = "None"
+    assert field_validator.validate("type", mocker.config_dict["type"]) == ALL
 
     field_validator.is_one_of = None
     field_validator.default = None
-    for key in json_dict.keys():
-        field_validator.is_instance = type(json_dict[key])
-        assert field_validator.validate(key, json_dict[key]) == json_dict[key]
-    json_dict["closed_on"] = None
-    assert field_validator.validate("closed_on", json_dict["closed_on"]) is None
+    for key in mocker.config_dict.keys():
+        field_validator.is_instance = type(mocker.config_dict[key])
+        assert (
+            field_validator.validate(key, mocker.config_dict[key])
+            == mocker.config_dict[key]
+        )
+    mocker.config_dict["closed_on"] = None
+    assert (
+        field_validator.validate("closed_on", mocker.config_dict["closed_on"]) is None
+    )
 
 
 def test_field_validator_validate_failure():
